Move SliceConfig to fv3fit._shared.config
