+++ external/fv3fit/fv3fit/_shared/__init__.py
@@ -1,6 +1,7 @@
 from .config import (
     TrainingConfig,
     register_training_function,
+    SliceConfig,
 )
 from .packer import (
     pack,
