+++ external/fv3fit/fv3fit/emulation/data/__init__.py
@@ -1,4 +1,4 @@
-from .config import TransformConfig, SliceConfig
+from .config import TransformConfig
 from . import transforms
 from .load import nc_files_to_tf_dataset, nc_dir_to_tf_dataset, batches_to_tf_dataset
 from .io import get_nc_files
